+++ app/Plugins/Core/src/Controller/IndexController.php
@@ -27,14 +27,14 @@ class IndexController
             ->with('tag', 'user')
             ->orderBy('topping', 'desc')
             ->orderBy('updated_at', 'desc')
-            ->paginate(get_options('topic_home_num', 15));
+            ->paginate((int)get_options('topic_home_num', 15));
         if (request()->input('query') === 'hot') {
             $page = Topic::query()
                 ->where('status', 'publish')
                 ->with('tag', 'user')
                 ->orderBy('view', 'desc')
                 ->orderBy('id', 'desc')
-                ->paginate(get_options('topic_home_num', 15));
+                ->paginate((int)get_options('topic_home_num', 15));
             $title = '热度最高的帖子';
         }
         if (request()->input('query') === 'publish') {
@@ -42,7 +42,7 @@ class IndexController
                 ->where('status', 'publish')
                 ->with('tag', 'user')
                 ->orderBy('id', 'desc')
-                ->paginate(get_options('topic_home_num', 15));
+                ->paginate((int)get_options('topic_home_num', 15));
             $title = '最新发布';
         }
         if (request()->input('query') === 'essence') {
@@ -50,7 +50,7 @@ class IndexController
                 ->where([['essence', '>', 0], ['status', 'publish']])
                 ->with('tag', 'user')
                 ->orderBy('updated_at', 'desc')
-                ->paginate(get_options('topic_home_num', 15));
+                ->paginate((int)get_options('topic_home_num', 15));
             $title = '精华';
         }
         if (request()->input('query') === 'topping') {
