+++ app/Plugins/Core/src/Controller/User/IndexController.php
@@ -84,7 +84,7 @@ class IndexController
             ->with('tag', 'user')
             ->orderBy('topping', 'desc')
             ->orderBy('id', 'desc')
-            ->paginate(get_options('topic_home_num', 15));
+            ->paginate((int)get_options('topic_home_num', 15));
 
         return view('User::draft', ['page' => $page]);
     }
